hub/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
google-genai>=1.0.0
httpx>=0.27.0
pydantic>=2.10.0
